debugedit: Add support for .debug_str_offsets (DW_FORM_strx)
In theory supporting strx .debug_str_offsets is easy, the strings in
.debug_str are just read through an indirection table. When the
strings are updated in .debug_str we just need to rewrite the
indirection table.

The tricky part is the ET_REL (object files or kernel modules)
support. Relocation reading is "global" per section and we expect to
read a relocation only once. But we need to read the
DW_AT_str_offsets_base before reading any strx form attributes. So we
read that first, then reset the relptr. And when we read from the
.debug_str_offsets section we need to save and restore the .debug_info
relptr.

	* tools/debugedit.c (do_read_24): New function.
	(str_offsets_base): New static variable.
	(buf_read_ule24): New function.
	(buf_read_ube24): Likewise.
	(setup_relbuf): Handle .debug_str_offsets.
	(do_read_uleb128): New function.
	(do_read_str_form_relocated): Likewise.
	(read_abbrev): Handle DW_FORM_strx[1234].
	(edit_strp): Take the actual string form as argument.
	Use do_read_str_form_relocated.
	(read_dwarf5_line_entries): Pass form to edit_strp.
	(edit_attributes_str_comp_dir): Take the actual string
	form as argument. Use do_read_str_form_relocated.
	(edit_attributes): Handle DW_FORM_strx[1234].
	(edit_info): Read DW_AT_str_offsets_base first.
	(update_str_offsets): New function.
	(edit_dwarf2): Setup do_read_24. Call update_str_offsets.

https://sourceware.org/bugzilla/show_bug.cgi?id=28728

Signed-off-by: Mark Wielaard <mark@klomp.org>