debugedit: Add -p, --preserve-dates
When --preserve-dates is given to debugedit it will take the original
st_atim and st_mtim and put them back using utimensat after processing
the file (and setting back the original st_mode).

Use it in find-debuginfo when processing ar archive entries so
timestamps in new archives are the same when put back.

    * tools/debugedit.c (preserve_dates): New bool.
    (optionsTable): Add preserve-dates.
    (optionsChars): Add p.
    (helpText): Add -p|--preserve-dates.
    (main): Handle 'p'. Use utimensat if preserve_dates true.
    * scripts/find-debuginfo.in (do_ar_file): Use debugedit -p.
    * tests/debugedit.at: Add new debugedit preserve timestamps test.
    * tests/find-debuginfo.at: Select the modes, date and entry names
    in the ar and compare them after find-debuginfo processing.

https://sourceware.org/bugzilla/show_bug.cgi?id=33096