find-debuginfo: Add -O/--output-dir switch
Originally, as it seems, the meaning of the builddir argument was that
of an *output* directory (for the list files), since that's all it was
used for. The actual *build* directory is pointed to by $RPM_BUILD_DIR.

However, since commit a8c7efeae67856e144897efb7c40516588b811ee, we have
been passing the builddir argument to "debugedit --base-dir" when unique
dirs are requested, effectively making the argument live up to its name.
Thus, the output directory currently equals the builddir and cannot be
set independently when needed. Fix that by adding a new switch to allow
just that.

This switch will be useful in RPM's %{__find_debuginfo} macro where we
can prevent "rpmbuild --build-in-place" from polluting the source tree
(since builddir points to $PWD in that mode).

Use -p and -o in the included test to also cover that code path.

Signed-off-by: Michal Domonkos <mdomonko@redhat.com>